core/requirements.txt
anchor-kr==0.1.3
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.34.154
botocore==1.34.154
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.1.14
cssbeautifier==1.15.1
cssselect==1.2.0
Django==5.0.7
django-allauth==64.0.0
djlint==1.34.1
EditorConfig==0.12.4
fake-useragent==1.5.1
feedparser==6.0.11
frozendict==2.4.4
gcloud==0.18.3
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-language==2.14.0
googleapis-common-protos==1.63.0
grpcio==1.65.4
grpcio-status==1.65.4
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
importlib_metadata==8.2.0
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
lxml==5.2.2
lxml_html_clean==0.2.0
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.4
multitasking==0.0.11
numpy==1.24.0
oauth2client==4.1.3
packaging==23.1
pandas==2.2.2
parse==1.20.2
pathspec==0.12.1
peewee==3.17.6
pkginfo==1.9.6
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.14.1
pyee==11.1.0
Pygments==2.16.1
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
regex==2023.12.25
requests==2.32.3
requests-html==0.10.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.10.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
w3lib==2.2.1
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yfinance==0.2.41
zipp==3.19.2
